requirements.txt
python-jobspy>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
scrapling[fetchers]>=0.3.14
camoufox>=0.4.0
python-docx>=1.1.0
